mixinを使って共通化
